+++ nampy/configs/__init__.py
@@ -10,7 +10,7 @@ from .namformer_config import DefaultNAMformerConfig
 from .natt_config import DefaultNATTConfig
 from .nbm_config import DefaultNBMConfig
 from .nodegam_config import DefaultNodeGAMConfig
-from .snam_config import DefaultSNAMConfig
+from .sparse_nam_config import DefaultSparseNAMConfig
 from .spline_nam_config import DefaultSplineNAMConfig
 
 # Note: GPNAM and QNAM use DefaultNAMConfig, so they don't have separate configs
@@ -22,7 +22,7 @@ __all__ = [
     "DefaultNATTConfig",
     "DefaultNAMformerConfig",
     "DefaultLinRegConfig",
-    "DefaultSNAMConfig",
+    "DefaultSparseNAMConfig",
     "DefaultSplineNAMConfig",
     "DefaultNodeGAMConfig",
 ]